updated readme, organized files
